backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
spacy==3.7.2
sentence-transformers==2.3.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
faiss-cpu==1.7.4
pdfminer.six==20231228
python-docx==1.1.0
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
